relocates single-use components
